+++ tests/features/treferences.nim
@@ -860,6 +860,85 @@ run_test("textDocument/references: module port (2)",
 )
 
 
+# Open the file "./src/src4.v", expecting no parsing errors.
+const src4_path = "./src/src4.v"
+const src4_text = static_read(src4_path)
+send(ifs, new_lsp_notification("textDocument/didOpen", %*{
+   "textDocument": {
+      "uri": "file://" & expand_filename(src4_path),
+      "languageId": "verilog",
+      "version": 0,
+      "text": src4_text
+   }
+}))
+assert len(recv(ofs).parameters["diagnostics"]) == 0
+
+const src5_path = "./src/src5.v"
+let src5_path_len = len(expand_filename(src5_path))
+
+run_test("textDocument/references: module instantiation",
+   new_lsp_request(0, "textDocument/references", %*{
+      "textDocument": {
+         "uri": "file://" & expand_filename(src3_path),
+      },
+      "position": {
+         "line": 23,
+         "character": 6
+      },
+      "context": {
+         "includeDeclaration": true
+      }
+   }),
+   new_lsp_response(227 + src3_path_len + src5_path_len, 0, %*[
+   {
+      "uri": "file://" & expand_filename(src5_path),
+      "range": {
+         "start": {"line": 15, "character": 12},
+         "end" : {"line": 15, "character": 24}
+      }
+   },
+   {
+      "uri": "file://" & expand_filename(src3_path),
+      "range": {
+         "start": {"line": 23, "character": 12},
+         "end" : {"line": 23, "character": 24}
+      }
+   }
+   ])
+)
+
+run_test("textDocument/references: module declaration",
+   new_lsp_request(0, "textDocument/references", %*{
+      "textDocument": {
+         "uri": "file://" & expand_filename(src4_path),
+      },
+      "position": {
+         "line": 0,
+         "character": 32
+      },
+      "context": {
+         "includeDeclaration": true
+      }
+   }),
+   new_lsp_response(227 + src3_path_len + src5_path_len, 0, %*[
+   {
+      "uri": "file://" & expand_filename(src5_path),
+      "range": {
+         "start": {"line": 15, "character": 12},
+         "end" : {"line": 15, "character": 24}
+      }
+   },
+   {
+      "uri": "file://" & expand_filename(src3_path),
+      "range": {
+         "start": {"line": 23, "character": 12},
+         "end" : {"line": 23, "character": 24}
+      }
+   }
+   ])
+)
+
+
 # Shut down the server.
 shutdown(ifs, ofs)